Fixed my problem generating the appcache manifest file:  the path to app.js and main.css are correct.
However, now that the manifest works, it exposed another problem:  the browser CACHES the manifest so that it doesn't see that the manifest has changed.

I could fix this on tacocat with some .htaccess changes, but there's no way to fix it on local build.

I'm disabling the appcache manifest in this commit until I get time to turn it on for just production builds, and modify the htaccess to ensure it doesn't get cached, and also ensure it gets served as the right mime type:  text/cache-manifest.